Remove redundant scripts

+++ scripts/setup_drone.bash
@@ -1,9 +1,8 @@
 #!/bin/bash
 set -e  # Exit on first error
 
-rm -rf $HOME/* # Clear all folders
+rm -rf "${HOME:?}"/* # Clear all folders
 sudo apt-get -qq install git make
 
 git clone https://github.com/chutsu/red
-cd red
-make install_deps
+cd red && make install_deps
